# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a clean and modular boilerplate for Go REST APIs with RBAC, Audit Logs, and WebSockets.",
        "title": "Go Clean Boilerplate API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/stats/activity": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns daily activity metrics (logins, audit events) for a given period.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get activity chart data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of days to retrieve (default 7)",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_pkg_response.SwaggerSuccessResponseWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_internal_modules_stats_model.DashboardActivity"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_pkg_response.SwaggerErrorResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_pkg_response.SwaggerErrorResponseWrapper"
                        }
                    }
                }
            }
        },
        "/stats/insights": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns system-level performance metrics (latency, error rate, uptime).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get system health insights",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_pkg_response.SwaggerSuccessResponseWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_internal_modules_stats_model.SystemInsights"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_pkg_response.SwaggerErrorResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_pkg_response.SwaggerErrorResponseWrapper"
                        }
                    }
                }
            }
        },
        "/stats/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns high-level system statistics (total users, roles, audit logs, etc.)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get dashboard summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_pkg_response.SwaggerSuccessResponseWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_internal_modules_stats_model.DashboardSummary"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_pkg_response.SwaggerErrorResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_pkg_response.SwaggerErrorResponseWrapper"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Roisfaozi_go-clean-boilerplate_internal_modules_stats_model.ActivityPoint": {
            "type": "object",
            "properties": {
                "audits": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "logins": {
                    "type": "integer"
                }
            }
        },
        "github_com_Roisfaozi_go-clean-boilerplate_internal_modules_stats_model.DashboardActivity": {
            "type": "object",
            "properties": {
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_internal_modules_stats_model.ActivityPoint"
                    }
                }
            }
        },
        "github_com_Roisfaozi_go-clean-boilerplate_internal_modules_stats_model.DashboardSummary": {
            "type": "object",
            "properties": {
                "total_audit_logs": {
                    "type": "integer"
                },
                "total_org_members": {
                    "type": "integer"
                },
                "total_roles": {
                    "type": "integer"
                },
                "total_users": {
                    "type": "integer"
                }
            }
        },
        "github_com_Roisfaozi_go-clean-boilerplate_internal_modules_stats_model.SystemInsights": {
            "type": "object",
            "properties": {
                "avg_latency_ms": {
                    "type": "number"
                },
                "error_rate": {
                    "type": "number"
                },
                "most_active_role": {
                    "type": "string"
                },
                "uptime": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Roisfaozi_go-clean-boilerplate_pkg_response.PageMetadata": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_item": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_Roisfaozi_go-clean-boilerplate_pkg_response.SwaggerErrorResponseWrapper": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Roisfaozi_go-clean-boilerplate_pkg_response.SwaggerSuccessResponseWrapper": {
            "type": "object",
            "properties": {
                "data": {},
                "paging": {
                    "$ref": "#/definitions/github_com_Roisfaozi_go-clean-boilerplate_pkg_response.PageMetadata"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Type 'Bearer ' followed by a space and the access token.\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}